requirements.txt
python-iprofile
django>=1.8
ipdb==0.8.1
ipython==4.1.1
twine==1.6.5
